cleanup in xtest
