+++ xtest/src/xml_file.rs
@@ -8,12 +8,31 @@ use crate::Metadata;
 pub struct XmlFile {
     pub name: String,
     pub xml_path: PathBuf,
-    pub metadata_path: PathBuf,
     pub metadata: Metadata,
+    pub metadata_path: PathBuf,
 }
 
 impl XmlFile {
     pub fn read_xml_file(&self) -> String {
         read_to_string(&self.xml_path).unwrap()
+    }
+}
+
+#[derive(Debug, Clone, Deserialize, Serialize, Eq, PartialEq)]
+#[serde(rename_all = "snake_case")]
+pub enum Syntax {
+    /// The XML file is expected to be parsed with success.
+    Good {},
+    /// The XML parser is expected to encounter an error at the given location.
+    Bad {
+        character_position: u64,
+        line: u64,
+        column: u64,
+    },
+}
+
+impl Default for Syntax {
+    fn default() -> Self {
+        Syntax::Good {}
     }
 }
